generate migrator file
